Update featured image tests to use collections api
